add error messgae when try you to add site num the exists

+++ myapp/views.py
@@ -5,6 +5,10 @@ from .models import Site
 
 from django.shortcuts import redirect, get_object_or_404
 from django.apps import apps
+from django.http import HttpResponse
+from django.views.decorators.csrf import csrf_exempt
+from urllib.parse import urlencode
+from django.contrib import messages
 
 def home(request):
     sites = Site.objects.all()
@@ -16,28 +20,23 @@ def sites(request):
         try:
             Site.objects.create(**site_data)
         except IntegrityError:
-            return render(request, 'sites.html', {
-                'sites': Site.objects.all(),
-                'headers': [field.name for field in Site._meta.fields],
-                'error': 'A site with this Site_num already exists.'
-            })
-        return redirect('sites')
+            messages.error(request, "A site with this number already exists.")
+            return redirect('sites')
     
     sites = Site.objects.all()
-    
     headers = [{'name': field.name, 'verbose_name': field.verbose_name} for field in Site._meta.fields]
-    
-    # if sites.exists():
-    #     headers = sites.first()._meta.fields
-    # else:
-    #     headers = []
-    
     return render(request, 'sites.html', {'sites': sites, 'headers': headers})
 
+@csrf_exempt
 def delete_object(request, model_name, object_id):
-    model = apps.get_model(app_label='myapp', model_name=model_name)
-    obj = get_object_or_404(model, SiteID=object_id)
-    obj.delete()
-    return redirect('sites') 
-    
+    if request.method == 'POST':
+        model = apps.get_model(app_label='myapp', model_name=model_name)
+        # Dynamically get the primary key field name
+        pk_field = model._meta.pk.name
+        obj = get_object_or_404(model, **{pk_field: object_id})
+        obj.delete()
+        # Get the redirect URL from the query parameters
+        redirect_url = request.GET.get('redirect_url', 'sites')
+        return redirect(redirect_url)  # Redirect to the specified URL
+    return HttpResponse(status=405)  # Method Not Allowed if not POST
